studentsystem/requirements.txt
arrow==1.2.3
asgiref==3.5.2
async-generator==1.10
attrs==20.3.0
binaryornot==0.4.4
certifi==2022.6.15
cffi==1.15.0
chardet==5.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
cookiecutter==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
distlib==0.3.4
Django==4.1.2
django-browser-reload==1.6.0
django-cors-headers==3.13.0
django-js-asset==2.0.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
filelock==3.7.1
gunicorn==20.1.0
h11==0.13.0
idna==3.3
itypes==1.2.0
jellyfish==0.9.0
Jinja2==3.1.2
jinja2-time==0.2.0
lib50==3.0.4
Markdown==3.3.7
markdown2==2.4.3
MarkupSafe==2.1.1
oauthlib==3.2.2
outcome==1.2.0
pexpect==4.8.0
Pillow==9.2.0
platformdirs==2.5.2
ptyprocess==0.7.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.4
PyYAML==5.4.1
requests==2.28.1
requests-oauthlib==1.3.1
selenium==4.4.0
six==1.16.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
submit50==3.1.1
termcolor==1.1.0
text-unidecode==1.3
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.15.1
whitenoise==6.2.0
wsproto==1.1.0
